Remove "Structural Element" from quotes
Once the term _Structural Element_ has been defined, then there's no need to include it within quotes. Before this patch, and ignoring plural variations, it would appear as:

 * "Structural Element"
 * 'Structural Element'
 * "Structural element"
 * Structural Element

This patch converts them all to the final format (title case without quotes), except for the occurrence in the definition, and the occurrence in the definition of a previous term.